requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pdfplumber
requests
pydantic
sentence-transformers
faiss-cpu
openai
PyMuPDF
numpy